Redesign redux state breaking down reducers
This implements a more polished version of the redux state. Entities are
stored in a specific part of the state, and anything that uses those
entities will use a reference to their id instead.
This commit introduces UI errors because the React components' prop
types have to be changed.
This work was planned in commit
fee1ab484afbd8228bdecd0eb6559ef2ce5035c9.